Wing yunnan performance

See merge request android_team/collect_android!1571

+++ feature/review-data/src/main/java/cn/thecover/media/feature/review_data/basic_widget/intent/ReviewDataIntent.kt
@@ -25,5 +25,5 @@ sealed class ReviewDataIntent {
     data object RefreshManuscriptDiffusion : ReviewDataIntent()
     data object LoadMoreManuscriptDiffusion : ReviewDataIntent()
 
-    data class EditManuscriptScore(val id: Int, val score: Int) : ReviewDataIntent()
+    data class EditManuscriptScore(val id: Int, val score: Double) : ReviewDataIntent()
 }
